requirements.txt
BeautifulSoup4
requests
pyyaml